ros2

+++ ros2/lib.sh
@@ -80,7 +80,10 @@ install_ros2_dep_pkgs() {
     xargs -a "$ROS2_DEP_PKGS_FILE" sudo apt-get install -s \
         | grep "^Inst" | awk '{print $2}' | LC_ALL=C sort -n \
         > "$ROS2_DEP_PKGS_TO_INSTALL_FILE"
-    xargs -a "$ROS2_DEP_PKGS_FILE" sudo apt-get install -y
+
+    if [[ -s "$ROS2_DEP_PKGS_FILE" ]]; then
+        xargs -a "$ROS2_DEP_PKGS_FILE" sudo apt-get install -y
+    fi
 }
 
 ################################################################################
